Coded a main menu w/ paw
A first draft of the main menu, in which buttons are indicated by a paw that can move up and down.  Buttons do not react, and effects of pushing them commented out (as in the example, brokenFile.  Graphics still primitive, but as images are all separate, they can be replaced.